requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
jinja2>=3.0.0
colorama>=0.4.4
rich>=13.0.0

# Code analysis tools
flake8>=6.0.0
pylint>=2.15.0
bandit>=1.7.5
safety>=2.3.0
semgrep>=1.45.0

# Language-specific analyzers
pycodestyle>=2.10.0
pyflakes>=3.0.0
mccabe>=0.7.0
pydocstyle>=6.3.0

# Security scanning
pip-audit>=2.6.0
vulndb>=0.1.0

# File handling and utilities
pathspec>=0.11.0
gitpython>=3.1.0
chardet>=5.0.0

# Web and reporting
weasyprint>=59.0
markdown>=3.5.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0